requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Cerberus==1.3.4
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.2
djangorestframework==3.13.1
idna==3.3
python-environ==0.4.54
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
